+++ hive_generator/lib/hive_generator.dart
@@ -1,8 +1,8 @@
 import 'package:build/build.dart';
-import 'package:hive_ce_generator/src/adapters_generator.dart';
-import 'package:hive_ce_generator/src/registrar_builder.dart';
-import 'package:hive_ce_generator/src/registrar_intermediate_builder.dart';
-import 'package:hive_ce_generator/src/type_adapter_generator.dart';
+import 'package:hive_ce_generator/src/generator/adapters_generator.dart';
+import 'package:hive_ce_generator/src/builder/registrar_builder.dart';
+import 'package:hive_ce_generator/src/builder/registrar_intermediate_builder.dart';
+import 'package:hive_ce_generator/src/generator/type_adapter_generator.dart';
 import 'package:source_gen/source_gen.dart';
 
 /// Builds Hive TypeAdapters
