+++ hive_generator/lib/src/adapter_builder/class_adapter_builder.dart
@@ -5,16 +5,16 @@ import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:collection/collection.dart';
 import 'package:hive_ce/hive.dart';
-import 'package:hive_ce_generator/src/builder.dart';
+import 'package:hive_ce_generator/src/adapter_builder/adapter_builder.dart';
 import 'package:hive_ce_generator/src/helper/helper.dart';
 import 'package:source_gen/source_gen.dart';
 
 import 'package:hive_ce_generator/src/helper/type_helper.dart';
 
 /// TODO: Document this!
-class ClassBuilder extends Builder {
+class ClassAdapterBuilder extends AdapterBuilder {
   /// TODO: Document this!
-  ClassBuilder(
+  ClassAdapterBuilder(
     super.cls,
     super.getters,
     super.setters,
