+++ hive_generator/lib/src/builder/registrar_builder.dart
