+++ hive_generator/lib/src/builder/registrar_intermediate_builder.dart
